refactor(#95): 포스트 서비스 분리
